Horizon login page contains DOS attack mechanism
the horizon login page (really the middleware) accesses the session
too early in the login process, which will create session records
in the session backend.  This is especially problematic when non-cookie
backends are used.

This one part of fix for CVE-2014-8124, the other is in
python-django-openstack-auth.

Resolves: rhbz#1169637

Change-Id: I9d2c40403fb9b0cfb512f2ff45397cbe0b050c71
Closes-Bug: 1394370
(cherry picked from commit e8a66a4d92ae259a5ef004cafad1809942c66596)
Reviewed-on: https://code.engineering.redhat.com/gerrit/38215
Reviewed-by: Julie Pichon <jpichon@redhat.com>
Tested-by: Julie Pichon <jpichon@redhat.com>